Disabled exception handling on the data project to make it consistent with our other libs.
Renamed EndianAdapter to ByteSwapPacker. Centralized the BytePacker to better report what this is. We could additionally simplify this by removing the use of a template trait class to specify the packer/unpacker. We'll decide once we have a few more examples.

Added read/advance overrun tests to memstorage along with tests.

Added a plat_crt as an exploration into ways of handling code so primitive and pervasive that scoping it to an additional namespace/class under secondpylon isnt productive. plat may either evolve into crt or crt may be split out into a separate project aggregating some of the smaller primitive projects.

Replaced strcpy with a safer version - we don't have *_s lib support on our platforms yet.